[FEAT] Hashtag 기능 구현 완료(API 명세 고려 하지 않음)

+++ src/main/java/uni/backend/domain/MainCategory.java
@@ -11,7 +11,7 @@ public class MainCategory {
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
     @Column(name = "main_category_id")
-    private Long mainCategoryId;
+    private Integer mainCategoryId;
 
     @ManyToOne
     @JoinColumn(name = "profile_id")
